Changed controller/action delimeter from '#' to empty space.
PURPOSE: To greatly increase flexibility of valid redirect URLS.
I wanted to redirect to a fragment url (i.e. http://site.com/dashboard#messages) and I couldn't because previous logic kept assuming I was trying to redirect to a controller  action because of the '#' character.

ALSO, I reduced the amount of work the router had to do to determine if it should treat a route as a redirection, because whatever route lacks a space character, IS an automatic redirection regardless of the presence or lack of a protocol i.e Presence or lack of http://, https://ftp:// etc. So now the route '/dashboard' will automatically be treated as a redirect as it rightly should and doesn't need to be 'http://site.com/dashboard'

REFERENCE: http://stackoverflow.com/questions/2849756/list-of-valid-characters-for-the-fragment-identifier-in-an-url